requirements.txt
anyascii==0.3.1
asgiref==3.5.2
beautifulsoup4==4.11.1
bleach==5.0.1
bootstrap-admin==0.4.4
captcha==0.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.6
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-advanced-filters==2.0.0
django-annoying==0.10.6
django-appconf==1.0.5
django-classy-tags==3.0.1
django-cms==3.11.0
django-crispy-forms==1.14.0
django-filter==22.1
django-form-builder==0.16.7
django-formfactory==0.2.3
django-forms-dynamic==1.0.0
django-formtools==2.3
django-import-export==2.9.0
django-ipware==4.0.2
django-jazzmin==2.5.0
django-jet-reboot==1.3.1
django-jsonstore==0.5.0
django-material==1.11.3
django-modelcluster==6.0
django-nested-admin==3.4.0
django-nested-inline==0.4.5
django-permissionedforms==0.1
django-sekizai==4.0.0
django-simple-tags==0.6.0
django-simplemde==0.1.4
django-summernote==0.8.20.0
django-taggit==3.0.0
django-treebeard==4.5.1
django-viewflow==1.11.0
djangocms-admin-style==3.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
Faker==14.2.0
filesig @ git+https://github.com/peppelinux/FileSignatureValidator.git@63b82b5df50f61376ad53896b6565be46bcb6361
gunicorn==20.1.0
hashids==1.3.1
html5lib==1.1
idna==3.4
import-export==0.2.67.dev6
jsonfield==3.1.0
l18n==2021.3
Markdown==3.4.1
MarkupPy==1.14
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.4.4
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.3
typing_extensions==4.3.0
tzdata==2022.2
Unidecode==1.3.4
urllib3==1.26.12
wagtail==4.0.2
webencodings==0.5.1
whitenoise==6.2.0
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0